+++ internal/game/player.go
@@ -2,13 +2,13 @@ package game
 
 // Player represents a player in the game
 type Player struct {
-	ID       int    `json:"id"`
-	Name     string `json:"name"`
-	Money    int    `json:"money"`
-	Income   int    `json:"income"`
-	Lives    int    `json:"lives"`
-	Latency  int64  `json:"latency"`
-	LastPing int64  `json:"lastPing"`
+	ID       int     `json:"id"`
+	Name     string  `json:"name"`
+	Money    float64 `json:"money"`
+	Income   float64 `json:"income"`
+	Lives    int     `json:"lives"`
+	Latency  int64   `json:"latency"`
+	LastPing int64   `json:"lastPing"`
 }
 
 // Implement Crud interface for Player
